cron_indexing_scheduler/requirements.txt
boto3==1.18.53
PyYAML==5.4.1
